requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.6
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.34.0
